FrameBuffer: Reorganize class hierarchy by making FrameBuffer abstract
